dom/base/test/jsmodules/test_importIntroType.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test introduction type of an imported module</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var introType;
  var extractIntroType;

  info('start');
  SimpleTest.waitForExplicitFinish();

  // eslint-disable-next-line no-unused-vars
  function testIntroductionType() {
    extractIntroType();
    ok(introType == "importedModule", 'Check introduction type');
    SimpleTest.finish();
  }
</script>
<iframe src="iframe_extractIntroType.html"></iframe>
<script type="module">
import "./module_extractIntroType.mjs";
</script>
<body onload='testIntroductionType()'></body>
